Fix auto-import of syntax
