+++ src/main/scala/zio/kafka/client/serde/package.scala
@@ -1,5 +1,3 @@
 package zio.kafka.client
 
-package object serde extends Serdes {
-  // TODO how to get the syntax available here? so clients can do import zio.kafka.client.serde._
-}
+package object serde extends Serdes with SerdeSyntax
